Comments in python.
Comments can be used to explain Python code. Comments can be used to make the code more readable. Comments can be used to prevent execution when testing code.
Comments start with a #, and Python will ignore them. Comments can be placed at the end of a line, and Python will ignore the rest of the line. A comment does not have to be text that explains the code, it can also be used to prevent Python from executing code.